refactor(staticfiles): split sidebar functions and styling into related files
