+++ addons/base/static/css/style.css
@@ -1,10 +1,10 @@
+@import "sidebar.css";
+
 /*--------------------------------------------------------------
 # General
 --------------------------------------------------------------*/
 :root {
   scroll-behavior: smooth;
-  --sidebar-width: 300px;
-  --sidebar-icons-width: 88px;
 }
 
 body {
@@ -587,223 +587,17 @@ h6 {
   background-color: #f6f9ff;
 }
 
-/*--------------------------------------------------------------
-# Sidebar
---------------------------------------------------------------*/
-.sidebar {
-  position: fixed;
-  top: 60px;
-  left: 0;
-  bottom: 0;
-  width: var(--sidebar-width);
-  z-index: 996;
-  transition: all 0.3s ease-in-out;
-  padding: 20px;
-  overflow-y: auto;
-  box-shadow: 0 0 20px rgba(1, 41, 112, 0.1);
-  background-color: #fff;
-  flex-direction: column;
-}
-
-/* Custom scrollbar for WebKit browsers (Chrome, Safari, Edge) */
-.sidebar::-webkit-scrollbar {
-  width: 6px; /* Adjust width */
-}
-
-.sidebar::-webkit-scrollbar-track {
-  background: transparent;
-}
-
-.sidebar::-webkit-scrollbar-thumb {
-  background-color: #aab7cf;
-  border-radius: 3px;
-}
-
-/* For Firefox (scrollbar-width is ignored in Safari) */
-@supports (scrollbar-width: thin) {
-  .sidebar {
-    scrollbar-width: thin;
-    scrollbar-color: #aab7cf transparent;
-  }
-}
-
-.sidebar.sidebar-open {
-  width: var(--sidebar-width);
-}
-
-.sidebar.sidebar-icons-only {
-  width: var(--sidebar-icons-width);
-}
-
-.sidebar.sidebar-close {
-  width: 0;
-  display: none;
-}
-
-.sidebar.sidebar-icons-only .sidebar-nav li a span {
-  display: none;
-}
-
-#main, #footer {
-  transition: margin-left 0.3s ease-in-out;
-  margin-left: var(--sidebar-width);
-}
-
-/* Full sidebar */
-.sidebar-open ~ #main,
-.sidebar-open ~ #footer {
-  margin-left: var(--sidebar-width);
-}
-
-/* Icons-only sidebar */
-.sidebar-icons-only ~ #main,
-.sidebar-icons-only ~ #footer {
-  margin-left: var(--sidebar-icons-width);
-}
-
-/* Sidebar hidden */
-.sidebar-close ~ #main,
-.sidebar-close ~ #footer {
-  margin-left: 0;
-}
-
 @media (max-width: 1199px) {
-  .sidebar {
-    display: none;
-  }
-
   #main,
   #footer {
     margin-left: 0 !important;
   }
-
-  .icon-sidebar-toggle {
-    display: none;
-  }
-}
-
-.sidebar::-webkit-scrollbar {
-  width: 5px;
-  height: 8px;
-  background-color: #fff;
-}
-
-.sidebar::-webkit-scrollbar-thumb {
-  background-color: #aab7cf;
 }
 
 @media (min-width: 500px) {
   .logo span {
     display: inline !important;
   }
-}
-
-.sidebar-nav {
-  padding: 0;
-  margin: 0;
-  list-style: none;
-  flex-grow: 1;
-}
-
-.sidebar-nav-bottom {
-  bottom: 20px;
-  position: absolute;
-  width: calc(100% - 40px);
-}
-
-.sidebar-nav li {
-  padding: 0;
-  margin: 0;
-  list-style: none;
-}
-
-.sidebar-nav .nav-item {
-  margin-bottom: 5px;
-}
-
-.sidebar-nav .nav-heading {
-  font-size: 11px;
-  text-transform: uppercase;
-  color: #899bbd;
-  font-weight: 600;
-  margin: 10px 0 5px 15px;
-}
-
-.sidebar-nav .nav-link {
-  display: flex;
-  align-items: center;
-  font-size: 15px;
-  font-weight: 600;
-  color: #4154f1;
-  transition: 0.3s;
-  background: #f6f9ff;
-  padding: 10px 15px;
-  border-radius: 4px;
-}
-
-.sidebar-nav .nav-link i {
-  font-size: 16px;
-  margin-right: 10px;
-  color: #4154f1;
-}
-
-.sidebar-nav .nav-link.collapsed {
-  color: #012970;
-  background: #fff;
-}
-
-.sidebar-nav .nav-link.collapsed i {
-  color: #899bbd;
-}
-
-.sidebar-nav .nav-link:hover {
-  color: #4154f1;
-  background: #f6f9ff;
-}
-
-.sidebar-nav .nav-link:hover i {
-  color: #4154f1;
-}
-
-.sidebar-nav .nav-link .bi-chevron-down {
-  margin-right: 0;
-  transition: transform 0.2s ease-in-out;
-}
-
-.sidebar-nav .nav-link:not(.collapsed) .bi-chevron-down {
-  transform: rotate(180deg);
-}
-
-.sidebar-nav .nav-content {
-  padding: 5px 0 0 0;
-  margin: 0;
-  list-style: none;
-}
-
-.sidebar-nav .nav-content a {
-  display: flex;
-  align-items: center;
-  font-size: 14px;
-  font-weight: 600;
-  color: #012970;
-  padding: 10px 0 10px 40px;
-  transition: 0.3s;
-}
-
-.sidebar-nav .nav-content a i {
-  font-size: 6px;
-  margin-right: 8px;
-  line-height: 0;
-  border-radius: 50%;
-}
-
-.sidebar-nav .nav-content a:hover,
-.sidebar-nav .nav-content a.active {
-  color: #4154f1;
-}
-
-.sidebar-nav .nav-content a.active i {
-  background-color: #4154f1;
 }
 
 /*--------------------------------------------------------------
