+++ addons/base/static/js/main.js
@@ -1,84 +1,10 @@
+import {select, on, onscroll} from "./helpers.js";
+import {Sidebar} from "./sidebar.js";
+
+Sidebar.init();
+
 (function() {
   "use strict";
-
-  /**
-   * Easy selector helper function
-   */
-  const select = (el, all = false) => {
-    el = el.trim()
-    if (all) {
-      return [...document.querySelectorAll(el)]
-    } else {
-      return document.querySelector(el)
-    }
-  }
-
-  /**
-   * Easy event listener function
-   */
-  const on = (type, el, listener, all = false) => {
-    if (all) {
-      select(el, all).forEach(e => e.addEventListener(type, listener))
-    } else {
-      select(el, all).addEventListener(type, listener)
-    }
-  }
-
-  /**
-   * Easy on scroll event listener 
-   */
-  const onscroll = (el, listener) => {
-    el.addEventListener('scroll', listener)
-  }
-
-  /**
-   * Toggle Sidebar States
-   */
-
-  const toggleSidebarState = () => {
-    // TODO: update to store preference to signed in user
-    let sidebarState = localStorage.getItem("sidebarState") || "open";
-
-    // Cycle through states: open -> icons-only -> close
-    if (sidebarState === "close") {
-      sidebarState = "open"
-    } else if (sidebarState === "open") {
-      sidebarState = "icons-only"
-    } else if (sidebarState === "icons-only") {
-      sidebarState = "close"
-    }
-    localStorage.setItem("sidebarState", sidebarState);
-    applySidebarClasses();
-  }
-
-  /**
-   * Apply Sidebar Classes
-   */
-
-  const applySidebarClasses = () => {
-    let sidebarState = localStorage.getItem("sidebarState") || "open";
-    const icon = select('.icon-sidebar-toggle');
-    const sidebar = select('.sidebar');
-    if (icon && sidebar) {
-      icon.classList.toggle("fa-bars-staggered", sidebarState === "open");
-      icon.classList.toggle("fa-bars", sidebarState === "icons-only");
-      icon.classList.toggle("fa-not-equal", sidebarState === "close");
-
-      sidebar.classList.toggle("sidebar-close", sidebarState === "close");
-      sidebar.classList.toggle("sidebar-open", sidebarState === "open");
-      sidebar.classList.toggle("sidebar-icons-only", sidebarState === "icons-only");
-    }
-  };
-
-  /**
-   * Click Sidebar Toggle Event
-   */
-
-  if (select('.icon-sidebar-toggle')) {
-    on('click', '.icon-sidebar-toggle', toggleSidebarState)
-  }
-
-  window.addEventListener('DOMContentLoaded', applySidebarClasses)
 
   /**
    * Search bar toggle
