+++ addons/base/templates/base.html
@@ -104,6 +104,6 @@
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js"></script>
 
   <!-- Template Main JS File -->
-  <script src="/static/js/main.js"></script>
+  <script type="module" src="/static/js/main.js"></script>
 </body>
 </html>
